Number guesses game - homework
Assigment: 
Required:

Write a program that picks a random number between 1 and 100. The user can then repeatedly guess a number, and the program will respond by saying whether the number is higher, lower, or correct (at which point the game ends).

Example output:

I have chosen a number between 1 and 100. Try to guess it!
What is your guess? => 66
The number is lower!
What is your guess? => 13
The number is higher!
What is your guess? => 42
Correct!
The yellow part is what the user types in.

Voluntary:

Make a new number guessing game where the roles are switched:

The computer has to guess a number the user thinks of. 

The user can answer “Yes”, “Higher” or “Lower”.