Blockchain P2P implementation. Group project
This code is part of our article that was written for 1st semester course "Basic project 1 – Applications of science in technology and society": 
**Blockchain based credentials**
How can blockchain technology improve the safety and
efficiency of storing academic credentials?


Motivation for this project:
Blockchain, initially proposed in 2008 as the technological foundation for Bitcoin, has gained tremendous popularity over the last decade and holds potential applications across various fields. Blockchain is often discussed as a technology capable of revolutionizing data storage and enhancing trust between society and institutions. This extends to areas such as medical records, digital identities, property records, and academic credentials. In light of the rising number of data breaches, it becomes imperative to explore secure data storage alternatives. Moreover, as the world grows increasingly interconnected, a centralized database for storing and easily verifying credentials could bring great convenience to students, employers, and institutions worldwide. Our objective is to investigate how blockchain technology could enhance the security and efficiency of academic credential storage and explore the possibilities, benefits, and risks associated with its use. In addition to literature analysis, we have developed a simplified blockchain prototype and a peer-to-peer network to illustrate the basic functioning of blockchain.

In recent years, blockchain technology has witnessed a surge in popularity, continuing its growth trajectory, and discovering new applications beyond cryptocurrencies. The number of projects within the blockchain domain has notably increased and continues to expand. Most blockchain-based initiatives and applications are geared toward enhancing bureaucratic accountability, transparency, and economic efficiency. Distributed ledger technology has gradually become a focal point of mainstream research and policy discussions. The enthusiastic response, driven by the success of Bitcoin and the emergence of new use cases, has raised high expectations for the revolutionary role of blockchain in business and the public sector. Despite the presence of hundreds of pilot projects, there are only a few notable government blockchain deployment success stories. This has piqued our curiosity to delve deeper into whether blockchain technology could offer solutions to longstanding issues or supplant outdated systems. Therefore, we aim to examine the potential applications of blockchain technology in the field of education and its capacity to enhance the efficiency and security of academic credential storage and sharing.